Fix clippy warnings

+++ src/terminal_size.rs
@@ -46,7 +46,7 @@ pub fn terminal_size(fileno: c_int) -> io::Result<(u16, u16)> {
         let mut size: TermSize = mem::zeroed();
 
         if ioctl(fileno, tiocgwinsz(), &mut size as *mut _) == 0 {
-            Ok((size.col as u16, size.row as u16))
+            Ok((size.col, size.row))
         } else {
             Err(io::Error::new(
                 io::ErrorKind::Other,
